Add day 6 - part 1 & part 2
Had some fun with this one. Indeed, part 1 was fairly simple reading the explanation. But it was actually not performant at all :D. So trying to scale it (with part 2) wasn’t working at all.

I had to rethink the concept in part2. I left part1 as it to see the difference :). Obviously applying part2 to part1 would be better.